6/16/16 by Ashwin
Added a MusicController, MusicEditor, ConsoleView, and MIDIView. These are the Controller, entry point to the program, and two Views, respectively. I changed Piece to remove the borders, and edited PieceTest to reflect that. Added the Util package, which has CompositionBuilder and its implementation, as well as MusicReader. Added SoundView, will delete though.